feat/ 새로 생긴 ParticipationStat에 맞추어, 내가 등록한 챌린지 보는 온보딩 페이지의 반환 DTO 관련 수정 #240

+++ src/main/java/com/habitpay/habitpay/domain/ChallengeAbsenceFee/exception/DaysCountException.java
@@ -5,7 +5,7 @@ import com.habitpay.habitpay.global.error.exception.IllegalArgumentException;
 
 public class DaysCountException extends IllegalArgumentException {
 
-    public DaysCountException(Long count) {
+    public DaysCountException(int count) {
         super(String.format("[%d] is illegal argument for TotalParticipatingDaysCount", count),
                 ErrorCode.TOTAL_PARTICIPATING_DAYS_COUNT_VALUE_IS_ILLEGAL);
     }
